@@ -11,7 +11,7 @@ namespace Nethermind.Core.Test.Json
 {
     public class ConverterTestBase<T>
     {
-        protected void TestConverter(T item, Func<T, T, bool> equalityComparer, JsonConverter<T> converter)
+        protected void TestConverter(T? item, Func<T, T, bool> equalityComparer, JsonConverter<T> converter)
         {
             JsonSerializer serializer = new();
             serializer.Converters.Add(converter);
