@@ -77,7 +77,9 @@ public class Eth67ProtocolHandlerTests
     [TearDown]
     public void TearDown()
     {
-        _handler.Dispose();
+        _handler?.Dispose();
+        _session?.Dispose();
+        _syncManager?.Dispose();
     }
 
     [Test]
