@@ -66,7 +66,7 @@ namespace Nethermind.Precompiles.Benchmark
                     {
                         EthereumJsonSerializer jsonSerializer = new EthereumJsonSerializer();
                         var jsonInputs = jsonSerializer.Deserialize<JsonInput[]>(File.ReadAllText(file));
-                        var parameters = jsonInputs.Select(i => new Param(precompile, i.Name, i.Input, i.Expected));
+                        var parameters = jsonInputs.Select(i => new Param(precompile, i.Name!, i.Input!, i.Expected));
                         inputs.AddRange(parameters);
                     }
 
